got things working. Just need to add descriptors
