--- src/main/resources/com/waytta/SaltAPIBuilder/local_batch.jelly
@@ -1,26 +1,0 @@
-<?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
-
-  <f:entry title="Batch Size" field="batchSize">
-    <f:textbox default="50%" />
-  </f:entry>
-  
-  <f:entry title="Target" field="target">
-    <f:textbox />
-  </f:entry>
-  
-  <f:entry title="Target Type" field="targetType">
-    <select name="targetType">
-      <f:option value="glob" selected="${instance.getTargettype() == 'glob'}">glob</f:option>
-      <f:option value="pcre" selected="${instance.getTargettype() == 'pcre'}">pcre</f:option>
-      <f:option value="list" selected="${instance.getTargettype() == 'list'}">list</f:option>
-      <f:option value="grain" selected="${instance.getTargettype() == 'grain'}">grain</f:option>
-      <f:option value="pillar" selected="${instance.getTargettype() == 'pillar'}">pillar</f:option>
-      <f:option value="nodegroup" selected="${instance.getTargettype() == 'nodegroup'}">nodegroup</f:option>
-      <f:option value="range" selected="${instance.getTargettype() == 'range'}">range</f:option>
-      <f:option value="compound" selected="${instance.getTargettype() == 'compound'}">compound</f:option>
-    </select>
-  </f:entry>
-
-  <!-- <f:class-entry clazz="com.waytta.clientinterface.BasicClient$LocalBatchClient" /> -->
-</j:jelly>
