+++ src/main/resources/com/waytta/clientinterface/LocalBatchClient/config.jelly
@@ -1,0 +1,26 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
+
+  <f:entry title="Batch Size" field="batchSize">
+    <f:textbox default="50%" />
+  </f:entry>
+  
+  <f:entry title="Target" field="target">
+    <f:textbox />
+  </f:entry>
+  
+  <f:entry title="Target Type" field="targetType">
+    <select name="targetType">
+      <f:option value="glob" selected="${instance.getTargetType() == 'glob'}">glob</f:option>
+      <f:option value="pcre" selected="${instance.getTargetType() == 'pcre'}">pcre</f:option>
+      <f:option value="list" selected="${instance.getTargetType() == 'list'}">list</f:option>
+      <f:option value="grain" selected="${instance.getTargetType() == 'grain'}">grain</f:option>
+      <f:option value="pillar" selected="${instance.getTargetType() == 'pillar'}">pillar</f:option>
+      <f:option value="nodegroup" selected="${instance.getTargetType() == 'nodegroup'}">nodegroup</f:option>
+      <f:option value="range" selected="${instance.getTargetType() == 'range'}">range</f:option>
+      <f:option value="compound" selected="${instance.getTargetType() == 'compound'}">compound</f:option>
+    </select>
+  </f:entry>
+
+  <!-- <f:class-entry clazz="com.waytta.clientinterface.BasicClient$LocalBatchClient" /> -->
+</j:jelly>
