+++ src/main/resources/com/waytta/clientinterface/LocalClient/config.jelly
@@ -2,7 +2,7 @@
 <j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
 
   <f:entry title="Wait for completion" field="blockbuild" description="Don't just send Salt command, wait for completion.">
-    <f:checkbox />
+    <f:checkbox/>
   </f:entry>
 
   <f:entry title="Poll Time" description="Number of seconds to wait before polling job completion status on jobs waiting for completion">
@@ -14,14 +14,14 @@
   </f:entry>
   <f:entry title="Target Type" field="targetType">
     <select name="targetType">
-      <f:option value="glob" selected="${instance.getTargettype() == 'glob'}">glob</f:option>
-      <f:option value="pcre" selected="${instance.getTargettype() == 'pcre'}">pcre</f:option>
-      <f:option value="list" selected="${instance.getTargettype() == 'list'}">list</f:option>
-      <f:option value="grain" selected="${instance.getTargettype() == 'grain'}">grain</f:option>
-      <f:option value="pillar" selected="${instance.getTargettype() == 'pillar'}">pillar</f:option>
-      <f:option value="nodegroup" selected="${instance.getTargettype() == 'nodegroup'}">nodegroup</f:option>
-      <f:option value="range" selected="${instance.getTargettype() == 'range'}">range</f:option>
-      <f:option value="compound" selected="${instance.getTargettype() == 'compound'}">compound</f:option>
+      <f:option value="glob" selected="${instance.getTargetType() == 'glob'}">glob</f:option>
+      <f:option value="pcre" selected="${instance.getTargetType() == 'pcre'}">pcre</f:option>
+      <f:option value="list" selected="${instance.getTargetType() == 'list'}">list</f:option>
+      <f:option value="grain" selected="${instance.getTargetType() == 'grain'}">grain</f:option>
+      <f:option value="pillar" selected="${instance.getTargetType() == 'pillar'}">pillar</f:option>
+      <f:option value="nodegroup" selected="${instance.getTargetType() == 'nodegroup'}">nodegroup</f:option>
+      <f:option value="range" selected="${instance.getTargetType() == 'range'}">range</f:option>
+      <f:option value="compound" selected="${instance.getTargetType() == 'compound'}">compound</f:option>
     </select>
   </f:entry>
