+++ src/main/resources/com/waytta/clientinterface/LocalClient/help-target.html
@@ -1,0 +1,3 @@
+<div>
+ Specify salt target as defined in the <a href=http://docs.saltstack.com/en/latest/topics/targeting/>salt documentation</a>. Paramaterized Build variables should be specified in the format {{VARNAME}}.
+</div>
